requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
easyocr==1.7.0
Pillow==10.0.1
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pandas==2.1.3
regex==2023.10.3
requests==2.31.0
huggingface-hub==0.19.4